refactor the homescreen to not have anything unrealistic
involves altering the rowLink and adding navigation environment variables/values that the view can use to determine what to do with the navigation elements. i have no real sense about whether this is a *good* way to do this, but it mostly works (it is a little iffy when going back to home, in terms of the timing of when the logo comes in and the back arrow disappears).